+++ pybricks/common.h
@@ -27,9 +27,12 @@
 
 void pb_package_pybricks_init(bool import_all);
 
-#if PYBRICKS_PY_COMMON_BLE
+#if PYBRICKS_PY_MESSAGING_BLE_RADIO
+extern const mp_obj_type_t pb_type_ble_radio;
+#endif
+
+#if PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD
 mp_obj_t pb_type_BLE_new(mp_obj_t broadcast_channel_in, mp_obj_t observe_channels_in);
-void pb_type_ble_start_cleanup(void);
 #endif
 
 #if PYBRICKS_PY_COMMON_CHARGER