+++ pybricks/hubs/pb_type_essentialhub.c
@@ -27,7 +27,7 @@
 typedef struct _hubs_EssentialHub_obj_t {
     mp_obj_base_t base;
     mp_obj_t battery;
-    #if PYBRICKS_PY_COMMON_BLE
+    #if PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD
     mp_obj_t ble;
     #endif
     mp_obj_t buttons;
@@ -41,7 +41,7 @@ static mp_obj_t hubs_EssentialHub_make_new(const mp_obj_type_t *type, size_t n_a
     PB_PARSE_ARGS_CLASS(n_args, n_kw, args,
         PB_ARG_DEFAULT_OBJ(top_side, pb_type_Axis_Z_obj),
         PB_ARG_DEFAULT_OBJ(front_side, pb_type_Axis_X_obj)
-        #if PYBRICKS_PY_COMMON_BLE
+        #if PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD
         , PB_ARG_DEFAULT_NONE(broadcast_channel)
         , PB_ARG_DEFAULT_OBJ(observe_channels, mp_const_empty_tuple_obj)
         #endif
@@ -49,7 +49,7 @@ static mp_obj_t hubs_EssentialHub_make_new(const mp_obj_type_t *type, size_t n_a
 
     hubs_EssentialHub_obj_t *self = mp_obj_malloc(hubs_EssentialHub_obj_t, type);
     self->battery = MP_OBJ_FROM_PTR(&pb_module_battery);
-    #if PYBRICKS_PY_COMMON_BLE
+    #if PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD
     self->ble = pb_type_BLE_new(broadcast_channel_in, observe_channels_in);
     #endif
     self->buttons = pb_type_Keypad_obj_new(MP_OBJ_FROM_PTR(self), pb_type_button_pressed_hub_single_button);
@@ -62,7 +62,7 @@ static mp_obj_t hubs_EssentialHub_make_new(const mp_obj_type_t *type, size_t n_a
 
 static const pb_attr_dict_entry_t hubs_EssentialHub_attr_dict[] = {
     PB_DEFINE_CONST_ATTR_RO(MP_QSTR_battery, hubs_EssentialHub_obj_t, battery),
-    #if PYBRICKS_PY_COMMON_BLE
+    #if PYBRICKS_PY_MESSAGING_BLE_RADIO_OLD
     PB_DEFINE_CONST_ATTR_RO(MP_QSTR_ble, hubs_EssentialHub_obj_t, ble),
     #endif
     PB_DEFINE_CONST_ATTR_RO(MP_QSTR_buttons, hubs_EssentialHub_obj_t, buttons),