+++ tests/bluetooth/virtual0.py
@@ -1,11 +1,11 @@
-from pybricks.hubs import VirtualHub
 from pybricks.tools import wait
+from pybricks.messaging import BLERadio
 
-hub = VirtualHub(broadcast_channel=123)
+radio = BLERadio(broadcast_channel=123)
 
 for i in range(99):
-    hub.ble.broadcast(["Hello", i])
+    radio.broadcast(["Hello", i])
     wait(100)
 
-hub.ble.broadcast(b"STOP")
+radio.broadcast(b"STOP")
 wait(500)